requirements.txt
pymodm
flask
pytest
pytest-pep8
pytest-cov
numpy
python-dateutil==2.7.2
gunicorn
flask-cors